django_news_recommender-main/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.3
idna==3.4
joblib==1.2.0
newsapi-python==0.2.6
nltk==3.7
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
stopwords==1.0.0
threadpoolctl==3.1.0
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
